requirements.txt
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
fastapi==0.95.2
h11==0.14.0
httpcore==0.17.3
httptools==0.7.1
httpx==0.24.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.11.1
lxml==6.0.2
MarkupSafe==3.0.3
openai==1.109.1
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
pydantic==1.10.24
pytest==7.4.4
python-docx==0.8.11
python-dotenv==1.2.1
python-multipart==0.0.6
python-pptx==0.6.23
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
starlette==0.27.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.22.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xlsxwriter==3.2.9
